react query refactoring
