+++ src/utils/queryCLient.ts
@@ -1,5 +1,5 @@
 import { QueryClient } from "react-query";
 const queryClient = new QueryClient({
-  defaultOptions: { queries: { staleTime: 25000 } },
+  defaultOptions: { queries: { staleTime: 25000, keepPreviousData: true } },
 });
 export default queryClient;
